corrected errors where NoneType objects were passed through
fixed:
error a crash where a None object was passed for validation in whichInput() - typically this occurred where a Viewer was passed as the output and the group node was attached to an input > 1 and inputs less than the attached input were not connected.

tidied up the codebase a bit

added:
documentation for functions lacking some
debugLog() - takes a message to display to the console. Requires debugFlag to be set to value other than None. None turns off debugLog print statements. Plan is for function to accept integer values corresponding to levels of granularity with log tracking and to display relevant log messages.

to come:
debugFlag - int values > 0 will correspond to levels of logging display
todos - still to be acted upon.